# Conflicts:
#	package.json
#	public/index.html
#	public/report.html
#	public/sw.js
#	public/version
#	src/components/sidebarLeft/tabs/language.ts
#	src/lib/langPack.ts
#	src/lib/mtproto/dcConfigurator.ts

+++ src/lib/cacheStorage.ts
@@ -131,9 +131,13 @@ export default class CacheStorageController {
     return Promise.resolve(fakeWriter);
   }
 
-  public static toggleStorage(enabled: boolean) {
-    return Promise.all(this.STORAGES.map(storage => {
+  public static toggleStorage(enabled: boolean, clearWrite: boolean) {
+    return Promise.all(this.STORAGES.map((storage) => {
       storage.useStorage = enabled;
+
+      if(!clearWrite) {
+        return;
+      }
       
       if(!enabled) {
         return storage.deleteAll();
@@ -141,7 +145,3 @@ export default class CacheStorageController {
     }));
   }
 }
-
-//const cacheStorage = new CacheStorageController(); 
-//MOUNT_CLASS_TO.cacheStorage = cacheStorage;
-//export default cacheStorage;
